portrait/pyproject.toml
[tool.poetry]
name = "portrait"
version = "0.1.0"
description = "The small brother of Ubuntu Landscape"
authors = ["George-Andrei Iosif <andrei.iosif@canonical.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.3"
python-pam = "^2.0.2"
six = "^1.16.0"
pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"

[tool.poetry.scripts]
portrait = 'portrait.app:main'

[tool.black]
line-length = 79
preview = true
include = '\.pyi?$'
extend-exclude = """
(
    ^/.venv/*
)
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
